Move predicates under `EntityWithLifecycle`
